requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

channels==4.3.2
channels-redis==4.3
daphne==4.1.0
redis==7.1

PyPDF2==3.0.1
pdfplumber==0.11.4

faiss-cpu==1.11.0.post1

google-genai==1.52.0

python-dotenv==1.0.1
numpy==2.2.0

